Remove unused disk related methods from disk.h and system.h
Moved class defined in dish. to system.h
